+++ src/components/notes/Note.tsx
@@ -1,7 +1,7 @@
 import { useNavigate } from "react-router-dom";
 import Modal from "../common/Modal";
 import MessageSendBox from "./MessageSendBox";
-import { INote } from "../../api/note";
+import { INote } from "../../store/note";
 
 import { useNote } from "../../hooks/useNote";
 import useModal from "../../hooks/useModal";
