+++ src/components/post/FindPostList.tsx
@@ -120,7 +120,7 @@ const InfoBox = styled.div`
 export default function FindPostList() {
   const navigate = useNavigate();
 
-  const { posts, isLoading, isError } = usePost();
+  const { posts, isLoading } = usePost();
   // console.log(posts);
 
   return isLoading ? (
